Improve API of the form list component
