+++ addon/routes/index.js
@@ -1,6 +1,5 @@
 import Route from "@ember/routing/route";
-import { RouteQueryManager } from "ember-apollo-client";
 
-export default Route.extend(RouteQueryManager, {
+export default Route.extend({
   model() {}
 });
