+++ addon/templates/index.hbs
@@ -1,1 +1,1 @@
-{{cfb-form-list data=data}}
+{{cfb-form-list}}
