+++ tests/integration/components/cfb-form-list-test.js
@@ -2,33 +2,18 @@ import { module, test } from "qunit";
 import { setupRenderingTest } from "ember-qunit";
 import { render, click } from "@ember/test-helpers";
 import hbs from "htmlbars-inline-precompile";
-import { task } from "ember-concurrency";
-import { defineProperty } from "@ember/object";
-import { run } from "@ember/runloop";
+import setupMirage from "ember-cli-mirage/test-support/setup-mirage";
 
 module("Integration | Component | cfb-form-list", function(hooks) {
   setupRenderingTest(hooks);
+  setupMirage(hooks);
 
   test("it renders blockless", async function(assert) {
     assert.expect(5);
 
-    defineProperty(
-      this,
-      "data",
-      task(function*() {
-        return yield [
-          { node: { id: 1, name: "Test 1" } },
-          { node: { id: 2, name: "Test 2" } },
-          { node: { id: 3, name: "Test 3" } },
-          { node: { id: 4, name: "Test 4" } },
-          { node: { id: 5, name: "Test 5" } }
-        ];
-      })
-    );
+    this.server.createList("form", 5);
 
-    run(() => this.get("data").perform());
-
-    await render(hbs`{{cfb-form-list data=data}}`);
+    await render(hbs`{{cfb-form-list}}`);
 
     assert.dom("table").exists();
     assert.dom("table > thead").exists();
@@ -40,17 +25,7 @@ module("Integration | Component | cfb-form-list", function(hooks) {
   test("it displays an empty state", async function(assert) {
     assert.expect(1);
 
-    defineProperty(
-      this,
-      "data",
-      task(function*() {
-        return yield [];
-      })
-    );
-
-    run(() => this.get("data").perform());
-
-    await render(hbs`{{cfb-form-list data=data}}`);
+    await render(hbs`{{cfb-form-list}}`);
 
     assert.dom("table > tbody > tr > td").hasText("No forms found");
   });
@@ -58,21 +33,11 @@ module("Integration | Component | cfb-form-list", function(hooks) {
   test("it can trigger editing of a row", async function(assert) {
     assert.expect(2);
 
+    this.server.create("form");
+
     this.set("on-edit-form", () => assert.step("edit-form"));
 
-    defineProperty(
-      this,
-      "data",
-      task(function*() {
-        return yield [{ node: { id: 1, name: "Test 1" } }];
-      })
-    );
-
-    run(() => this.get("data").perform());
-
-    await render(
-      hbs`{{cfb-form-list data=data on-edit-form=(action on-edit-form)}}`
-    );
+    await render(hbs`{{cfb-form-list on-edit-form=(action on-edit-form)}}`);
 
     await click("table > tbody > tr:first-of-type");
 
@@ -82,24 +47,10 @@ module("Integration | Component | cfb-form-list", function(hooks) {
   test("it renders block style", async function(assert) {
     assert.expect(3);
 
-    defineProperty(
-      this,
-      "data",
-      task(function*() {
-        return yield [
-          { node: { id: 1, name: "Test 1" } },
-          { node: { id: 2, name: "Test 2" } },
-          { node: { id: 3, name: "Test 3" } },
-          { node: { id: 4, name: "Test 4" } },
-          { node: { id: 5, name: "Test 5" } }
-        ];
-      })
-    );
-
-    run(() => this.get("data").perform());
+    this.server.createList("form", 5);
 
     await render(hbs`
-      {{#cfb-form-list data=data as |table|}}
+      {{#cfb-form-list as |table|}}
         {{#table.head}}
           <tr>
             <th>Key 1</th>
